Initial Commit
The schematic is structured and ready. the footprints have been assigned. And the preliminary position of the components on the PCB has been established. 

However, there is a footprint reference to one of the components by a path in the local machine that may not work at first (to be corrected). One of the components on the PCB did not have its connections recognized (to be fixed)